fix(wire): SpiceString flags is u8 on the wire, not u16
spice.proto declares string_flags as flags8 (3-byte String header:
length u16 + flags u8), not u16 as the plan's protocol table had it —
every RasterGlyph after the header was one byte out. Also pins the
path/str nonnull-pointer guards and the num_segments/glyph-dimension
hostile-length caps with dedicated tests.

+++ Sources/SpiceWire/Geometry.swift
@@ -113,7 +113,7 @@ public struct SpiceString: Sendable, Equatable {
     public init(flags: UInt16, glyphs: [RasterGlyph]) { self.flags = flags; self.glyphs = glyphs }
     public init(reader r: inout SpiceReader) throws {
         let length = try r.u16()
-        flags = try r.u16()
+        flags = UInt16(try r.u8())
         let bpp = flags & StringFlags.rasterA8 != 0 ? 8 : flags & StringFlags.rasterA4 != 0 ? 4 : 1
         glyphs = try (0 ..< length).map { _ in
             let renderPos = try SpicePoint(reader: &r), origin = try SpicePoint(reader: &r)
